Moved Grid classes to separate file. Refreshing heatmap on events and drawing from historical measurements
