+++ app/src/main/java/io/lunarlogic/aircasting/screens/session_view/map/MapContainer.kt
@@ -128,6 +128,7 @@ class MapContainer: OnMapReadyCallback {
             drawFixedMeasurement()
         }
         // sometimes onMapReady is invoked earlier than bindStream
+        println("MARYSIA: map status in bindSession ${status.get()}")
         if (status.get() == Status.MAP_LOADED.value) {
             setup()
         }
@@ -193,7 +194,7 @@ class MapContainer: OnMapReadyCallback {
         val mapHeight = mMapFragment?.view?.height ?: 0
 
 
-        mMap?.let {map ->
+        mMap?.let { map ->
             val sensorThreshold = mSessionPresenter?.selectedSensorThreshold()
             sensorThreshold?.let { sensorThreshold ->
                 if (mMapGrid != null) {
@@ -201,9 +202,9 @@ class MapContainer: OnMapReadyCallback {
                     mMapGrid = null
                 }
                 mMapGrid = MapGrid(mContext, map, sensorThreshold, mapWidth, mapHeight )
+                println("MARYSIA: drawing heatmap mMeasurements count ${mMeasurements.size}")
+                mMapGrid?.drawMap(mMeasurements)
             }
-
-
         }
 
     }
@@ -313,19 +314,10 @@ class MapContainer: OnMapReadyCallback {
         return ColorPoint(point, color)
     }
 
-    private fun measurementSpan(measurement: Measurement) : StyleSpan {
-        if (measurement.latitude == null || measurement.longitude == null) return FALLBACK_SPAN
-        val threshold = mSessionPresenter?.selectedSensorThreshold() ?: return FALLBACK_SPAN
-        return when (val level = measurement.getLevel(threshold)) {
-            Measurement.Level.EXTREMELY_LOW -> FALLBACK_SPAN
-            Measurement.Level.EXTREMELY_HIGH -> FALLBACK_SPAN
-            else -> LEVEL_SPANS[level.value]
-        }
-    }
-
     fun refresh(sessionPresenter: SessionPresenter?) {
         clearMap()
         bindSession(sessionPresenter)
+        drawHeatMap()
         drawSession()
     }
 
@@ -365,194 +357,3 @@ class MapContainer: OnMapReadyCallback {
 }
 
 data class ColorPoint(val point: LatLng, val color: Int)
-
-class GridSquare {
-    private var mGridMap: MapGrid
-    private var mBounds: LatLngBounds
-    private var sum: Double = 0.0
-    private var number: Int = 0
-    private var averagedValue: Double = 0.0
-    private var level: Measurement.Level? = null
-    var mPolygonOptions: PolygonOptions
-    private var mPolygon: Polygon? = null
-    var newColor: Boolean = false
-    val mSouthWestLatLng: LatLng
-    val mSouthEastLatLng: LatLng
-    val mNorthEastLatLng: LatLng
-    val mNorthWestLatLng: LatLng
-
-    constructor(mapGrid: MapGrid, southWestLatLng: LatLng, southEastLatLng: LatLng, northEastLatLng: LatLng, northWestLatLng: LatLng) {
-        mGridMap = mapGrid
-        mBounds = LatLngBounds(southWestLatLng, northEastLatLng)
-        mSouthWestLatLng = southWestLatLng
-        mSouthEastLatLng = southEastLatLng
-        mNorthEastLatLng = northEastLatLng
-        mNorthWestLatLng = northWestLatLng
-
-        mPolygonOptions = PolygonOptions()
-            .add(southWestLatLng, southEastLatLng, northEastLatLng, northWestLatLng)
-            .strokeWidth(0f)
-    }
-
-    fun addMeasurement(measurement: Measurement) {
-        sum += measurement.value
-        number += 1
-        calculateAverage()
-        if (averagedValue > 0) {
-            val newLevel = Measurement.getLevel(averagedValue, mGridMap.sensorThreshold)
-            if (newLevel != level) {
-                level = newLevel
-                newColor = true
-                mPolygonOptions = mPolygonOptions.fillColor(getColor())
-            }
-
-        }
-
-    }
-
-    private fun getColor(): Int {
-        if (level == null) {
-            return Color.TRANSPARENT
-        } else {
-            return getTransparentColor(MeasurementColor.colorForLevel(mGridMap.context, level!!))
-        }
-    }
-
-    private fun getTransparentColor(color: Int) : Int{
-        var alpha = Color.alpha(color)
-        val red = Color.red(color);
-        val green = Color.green(color);
-        val blue = Color.blue(color);
-
-        alpha += 100;
-
-        return Color.argb(alpha, red, green, blue)
-    }
-    fun addPolygon(polygon: Polygon) {
-        println("MARYSIA: adding polygon ${polygon}")
-        if (mPolygon != null) {
-            mPolygon?.remove()
-        }
-        mPolygon = polygon
-        println("MARYSIA: polygon added? ${mPolygon}")
-        newColor = false
-    }
-
-    fun remove() {
-        println("MARYSIA: remove polygon ${mPolygon}")
-        mPolygon?.remove()
-    }
-    fun inBounds(coordinates: LatLng) : Boolean {
-        return mBounds.contains(coordinates)
-    }
-
-    private fun calculateAverage() {
-        averagedValue = sum / number
-    }
-
-
-}
-
-class MapGrid(val context: Context?, val map: GoogleMap, val sensorThreshold: SensorThreshold,  val mapWidth: Int, val mapHeight: Int) {
-    private val DENSITY = 10
-
-    var mGridSquares: HashMap<String, GridSquare?> = HashMap()
-    var latNorth: Double
-    var latSouth: Double
-    var lonEast: Double
-    var lonWest: Double
-    var gridSizeX: Int
-    var gridSizeY: Int
-
-    init {
-        val visibleRegion = map.projection.visibleRegion
-        latNorth = visibleRegion.farLeft.latitude
-        latSouth = visibleRegion.nearLeft.latitude
-        lonEast = visibleRegion.farRight.longitude
-        lonWest = visibleRegion.nearLeft.longitude
-
-        val size = Math.min(mapWidth, mapHeight) / DENSITY
-
-        gridSizeX = mapWidth / size
-        gridSizeY = mapHeight / size
-
-        val lonGridSide = (lonEast - lonWest) / gridSizeX
-        val latGridSide = (latNorth - latSouth) / gridSizeY
-
-
-        for (x in 1..gridSizeX) {
-            for (y in 1..gridSizeY) {
-                val southWestLatLng =
-                    LatLng(latSouth + (y - 1) * latGridSide, lonWest + (x - 1) * lonGridSide)
-                val southEastLatLng =
-                    LatLng(latSouth + (y - 1) * latGridSide, lonWest + x * lonGridSide)
-                val northEastLatLng = LatLng(latSouth + y * latGridSide, lonWest + x * lonGridSide)
-                val northWestLatLng =
-                    LatLng((latSouth + y * latGridSide), lonWest + (x - 1) * lonGridSide)
-                println("MARYSIA: southWestLatLng: ${southWestLatLng}")
-                mGridSquares["${x}_${y}"] = GridSquare(
-                    this,
-                    southWestLatLng,
-                    southEastLatLng,
-                    northEastLatLng,
-                    northWestLatLng
-                )
-            }
-        }
-    }
-
-        fun addMeasurement(measurement: Measurement) {
-            if (measurement.latitude == null || measurement.longitude == null) return
-
-            for (x in 1..gridSizeX) {
-                for (y in 1..gridSizeY) {
-                    val gridSquare = mGridSquares["${x}_${y}"]
-                    gridSquare?.let {
-                        if (gridSquare?.inBounds(
-                                LatLng(
-                                    measurement.latitude,
-                                    measurement.longitude
-                                )
-                            )
-                        ) {
-                            gridSquare.addMeasurement(measurement)
-                        }
-                    }
-
-                }
-            }
-
-            drawAverages()
-        }
-
-        private fun drawAverages() {
-            for (x in 1..gridSizeX) {
-                for (y in 1..gridSizeY) {
-
-                    val gridSquare = mGridSquares["${x}_${y}"]
-                    gridSquare?.let {
-                        if (gridSquare.newColor) {
-                            val polygon = map.addPolygon(gridSquare.mPolygonOptions)
-                            println("MARYSIA: x_y ${x}_${y}")
-                            gridSquare.addPolygon(polygon)
-                        }
-                    }
-
-                }
-            }
-        }
-
-        fun remove() {
-            for (x in 1..gridSizeX) {
-                for (y in 1..gridSizeY) {
-                    println("MARYSIA: x_y ${x}_${y}")
-                    mGridSquares["${x}_${y}"]?.remove()
-                    mGridSquares["${x}_${y}"] = null
-                }
-            }
-        }
-
-
-
-//    }
-}
